requirements.txt
scikit-learn==1.5.0
opencv-python==4.10.0.84
numpy==1.26.4
flask==3.0.3
pillow==10.4.0
scikit-image==0.23.2